Adding the drawer page and some minor improvings
