+++ Trabalho 01/.dart_tool/package_config.json
@@ -140,7 +140,7 @@
       "languageVersion": "2.7"
     }
   ],
-  "generated": "2020-10-19T02:53:20.782772Z",
+  "generated": "2020-10-21T00:24:59.323659Z",
   "generator": "pub",
   "generatorVersion": "2.9.2"
 }
