+++ Trabalho 01/.packages
@@ -1,4 +1,4 @@
-# Generated by pub on 2020-10-18 23:53:20.726683.
+# Generated by pub on 2020-10-20 21:24:59.138785.
 async:file:///Users/macbookpro/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.4.2/lib/
 boolean_selector:file:///Users/macbookpro/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.0.0/lib/
 characters:file:///Users/macbookpro/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.0.0/lib/
